requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
anthropic==0.43.0
python-dotenv==1.0.1
python-multipart==0.0.20
watchdog==6.0.0
